Manager: Move settings to the Configuration namespace
