+++ ArcdpsLogManager/Dialogs/ApiDialog.cs
@@ -2,6 +2,7 @@ using System;
 using System.IO;
 using Eto.Drawing;
 using Eto.Forms;
+using GW2Scratch.ArcdpsLogManager.Configuration;
 using GW2Scratch.ArcdpsLogManager.Gw2Api;
 using GW2Scratch.ArcdpsLogManager.Processing;
