+++ ArcdpsLogManager/ManagerForm.cs
@@ -9,6 +9,7 @@ using Eto.Forms;
 using GW2Scratch.ArcdpsLogManager.Analytics;
 using GW2Scratch.ArcdpsLogManager.Collections;
 using GW2Scratch.ArcdpsLogManager.Commands;
+using GW2Scratch.ArcdpsLogManager.Configuration;
 using GW2Scratch.ArcdpsLogManager.Controls;
 using GW2Scratch.ArcdpsLogManager.Controls.Filters;
 using GW2Scratch.ArcdpsLogManager.Dialogs;
